feat: download trasfersdk automatically for make
It's probably not a good idea to include the aspera transfer sdk itself
to this code base, so we need a way to automatically configuring the sdk
including downloading the file, modifying the proto file, generating the
pb go code and placing them to the right place.

With this commit, dev can do it with one `make` command.

fixes: #7